gh-138122: Implement PEP 799
