--- Lib/profile/__init__.py
@@ -1,6 +1,0 @@
-from .profile import run
-from .profile import runctx
-from .profile import Profile
-from .profile import _Utils
-
-__all__ = ['run', 'runctx', 'Profile']
