+++ Lib/profiling/sampling/__main__.py
@@ -1,0 +1,6 @@
+"""Run the sampling profiler from the command line."""
+
+from .sample import main
+
+if __name__ == '__main__':
+    main()
