+++ Lib/profiling/sampling/collector.py
