+++ Lib/profiling/sampling/pstats_collector.py
