+++ Lib/profiling/sampling/stack_collector.py
