+++ Lib/profiling/tracing/__main__.py
@@ -1,0 +1,6 @@
+"""Run the tracing profiler from the command line."""
+
+from . import main
+
+if __name__ == '__main__':
+    main()
